Extract URL into a separate module
