+++ setup.cfg
@@ -24,5 +24,9 @@ max-line-length = 88
 # warn_redundant_casts = True
 # warn_unused_ignores = True
 
+[mypy-idna]
+ignore_missing_imports = true
+
 [mypy-pytest]
 ignore_missing_imports = true
+
